ai-engine/requirements.txt
fastapi
uvicorn
yfinance
pandas
numpy
ta
xgboost
scikit-learn
ccxt
python-dotenv
